fix: move event modal to page component
