+++ src/app/evenements/page.tsx
@@ -1,12 +1,16 @@
-import Section from '@/components/Section'
 import { getBalEvents } from '@/lib/api-bal-admin'
 import banEvents from '@/data/ban-events.json'
-import CardWrapper from '@/components/CardWrapper'
-import EventCard from '@/components/Events/EventCard'
 import { EventType } from '@/types/events.types'
 import { getUpcomingAndPassedEvents, mapEvents } from '@/utils/events'
+import { createModal } from '@codegouvfr/react-dsfr/Modal'
+import EventPage from '@/components/Events/EventPage'
 
 export const dynamic = 'force-dynamic'
+
+const Modal = createModal({
+  id: 'register-event-modal',
+  isOpenedByDefault: false,
+})
 
 export default async function EvenementsPage() {
   const balEvents = await getBalEvents()
@@ -21,34 +25,6 @@ export default async function EvenementsPage() {
   })
 
   return (
-    <>
-      <Section title="Prochains évènements">
-        <CardWrapper>
-          {upcomingEvents.length > 0
-            ? upcomingEvents.map((event, index) => (
-              <EventCard
-                key={index}
-                event={event}
-                tagToColor={tagToColor}
-              />
-            ))
-            : <p>Aucun évènement à venir</p>}
-        </CardWrapper>
-      </Section>
-      <Section title="Évènements passés">
-        <CardWrapper>
-          {lastMonthPastEvents.length > 0
-            ? lastMonthPastEvents.map((event, index) => (
-              <EventCard
-                key={index}
-                event={event}
-                tagToColor={tagToColor}
-                isPassed
-              />
-            ))
-            : <p>Aucun évènement n&apos;a eu lieu le mois dernier</p>}
-        </CardWrapper>
-      </Section>
-    </>
+    <EventPage upcomingEvents={upcomingEvents} lastMonthPastEvents={lastMonthPastEvents} tagToColor={tagToColor} />
   )
 }
